Implement multi-cell checking
This addresses #5, adding an additional method on the DeciderImpl
trait and 2 more on the Decider trait, accomodating checks for n
packets.

This is a bit more difficult to reason about with GCRA, but still
possible: If you process 5 cells now, that uses up the "expected
arrival time" budget for the next 5 cells, pushing back the next
expected arrival time by 5*as much as a single cell would.