Only CAS in the positive case
In the negative case, we can get away with not writing at all - at the
current fill level, even if we retried, the decision would still be
negative. The only thing that *might* change is the wait period, but
that's a minimum anyway, so we can simply punt early for No's.